backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.31.0
sqlalchemy>=2.0.0,<2.1.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.9,<0.1.0
pydantic[email]>=2.7.0,<3.0.0
httpx>=0.27.0,<0.28.0
Pillow>=10.0.0,<11.0.0
inference-sdk>=0.9.0,<0.10.0
python-dotenv>=1.0.0,<2.0.0
